backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
pandas==2.2.1
requests==2.31.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
sqlalchemy==2.0.29
gunicorn==21.2.0
